+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - fix ripgrep link in README.md (thanks to [pke](https://github.com/pke))
 - Remove extension.js.map from bundle
+- Add context menu options to toggle badges and item counts
 
 ## v0.0.182 - 2020-10-03